Consolidate HeapUsage class and remove vmStructs_dd*
