+++ CLAUDE.md
@@ -213,7 +213,6 @@ The profiler uses a sophisticated double-buffered storage system for call traces
 - **Buffer Management**: Thread-local recording buffers with configurable flush thresholds
 
 ### Native Integration Patterns
-- **Adapter Pattern**: `*_dd.h` files adapt upstream code for Datadog needs
 - **Signal Handler Safety**: Careful memory management in signal handler contexts
 
 ### Multi-Engine Profiling System
