+++ ddprof-lib/src/main/cpp/vmStructs.cpp
@@ -8,6 +8,8 @@
 #include "vmStructs.h"
 #include "vmEntry.h"
 #include "j9Ext.h"
+#include "jniHelper.h"
+#include "jvmHeap.h"
 #include "safeAccess.h"
 #include "spinLock.h"
 
@@ -1083,4 +1085,134 @@ int VMThread::state() {
         }
     }
     return 0;
+}
+
+bool HeapUsage::is_jmx_attempted = false;
+bool HeapUsage::is_jmx_supported = false; // default to not-supported
+
+void HeapUsage::initJMXUsage(JNIEnv *env) {
+    if (is_jmx_attempted) {
+        // do not re-run the initialization
+        return;
+    }
+    is_jmx_attempted = true;
+    jclass factory = env->FindClass("java/lang/management/ManagementFactory");
+    if (!jniExceptionCheck(env) || factory == nullptr) {
+        return;
+    }
+    jclass memoryBeanClass = env->FindClass("java/lang/management/MemoryMXBean");
+    if (!jniExceptionCheck(env) || memoryBeanClass == nullptr) {
+        return;
+    }
+    jmethodID get_memory = env->GetStaticMethodID(
+        factory, "getMemoryMXBean", "()Ljava/lang/management/MemoryMXBean;");
+    if (!jniExceptionCheck(env) || get_memory == nullptr) {
+        return;
+    }
+    jobject memoryBean = env->CallStaticObjectMethod(factory, get_memory);
+    if (!jniExceptionCheck(env) || memoryBean == nullptr) {
+        return;
+    }
+    jmethodID get_heap = env->GetMethodID(memoryBeanClass, "getHeapMemoryUsage",
+                                          "()Ljava/lang/management/MemoryUsage;");
+    if (!jniExceptionCheck(env) || get_heap == nullptr) {
+        return;
+    }
+    env->CallObjectMethod(memoryBean, get_heap);
+    if (!jniExceptionCheck(env)) {
+        return;
+    }
+    // mark JMX as supported only after we were able to retrieve the memory usage
+    is_jmx_supported = true;
+}
+
+bool HeapUsage::isLastGCUsageSupported() {
+    // only supported for JDK 17+
+    // the CollectedHeap structure is vastly different in JDK 11 and earlier so
+    // we can't support it
+    return _collected_heap_addr != NULL && _heap_usage_func != NULL;
+}
+
+bool HeapUsage::needsNativeBindingInterception() {
+    return _collected_heap_addr == NULL ||
+           (_heap_usage_func == NULL && _gc_heap_summary_func == NULL);
+}
+
+jlong HeapUsage::getMaxHeap(JNIEnv *env) {
+    static jclass _rt;
+    static jmethodID _get_rt;
+    static jmethodID _max_memory;
+
+    if (!(_rt = env->FindClass("java/lang/Runtime"))) {
+        jniExceptionCheck(env);
+        return -1;
+    }
+
+    if (!(_get_rt = env->GetStaticMethodID(_rt, "getRuntime",
+                                          "()Ljava/lang/Runtime;"))) {
+        jniExceptionCheck(env);
+        return -1;
+    }
+
+    if (!(_max_memory = env->GetMethodID(_rt, "maxMemory", "()J"))) {
+        jniExceptionCheck(env);
+        return -1;
+    }
+
+    jobject rt = (jobject)env->CallStaticObjectMethod(_rt, _get_rt);
+    jlong ret = (jlong)env->CallLongMethod(rt, _max_memory);
+    if (jniExceptionCheck(env)) {
+        return -1;
+    }
+    return ret;
+}
+
+HeapUsage HeapUsage::get() {
+    return get(true);
+}
+
+HeapUsage HeapUsage::get(bool allow_jmx) {
+    HeapUsage usage;
+    if (_collected_heap_addr != NULL) {
+        if (_heap_usage_func != NULL) {
+            // this is the JDK 17+ path
+            usage = _heap_usage_func(*_collected_heap_addr);
+            usage._used_at_last_gc =
+                ((CollectedHeapWrapper *)*_collected_heap_addr)->_used_at_last_gc;
+        } else if (_gc_heap_summary_func != NULL) {
+            // this is the JDK 11 path
+            // we need to collect GCHeapSummary information first
+            GCHeapSummary summary = _gc_heap_summary_func(*_collected_heap_addr);
+            usage._initSize = -1;
+            usage._used = summary.used();
+            usage._committed = -1;
+            usage._maxSize = summary.maxSize();
+        }
+    }
+    if (usage._maxSize == -1 && _memory_usage_func != NULL && allow_jmx && isJMXSupported()) {
+        // this path is for non-hotspot JVMs
+        // we need to patch the native method binding for JMX GetMemoryUsage to
+        // capture the native method pointer first also, it requires JMX and
+        // allocating new objects so it really should not be used in a GC callback
+        JNIEnv *env = VM::jni();
+        if (env == NULL) {
+            return usage;
+        }
+        jobject m_usage =
+            (jobject)_memory_usage_func(env, (jobject)NULL, (jboolean) true);
+        jclass cls = env->GetObjectClass(m_usage);
+        jfieldID init_fid = env->GetFieldID(cls, "init", "J");
+        jfieldID max_fid = env->GetFieldID(cls, "max", "J");
+        jfieldID used_fid = env->GetFieldID(cls, "used", "J");
+        jfieldID committed_fid = env->GetFieldID(cls, "committed", "J");
+        if (init_fid == NULL || max_fid == NULL || used_fid == NULL ||
+            committed_fid == NULL) {
+            return usage;
+        }
+        usage._initSize = env->GetLongField(m_usage, init_fid);
+        usage._maxSize = env->GetLongField(m_usage, max_fid);
+        usage._used = env->GetLongField(m_usage, used_fid);
+        usage._committed = env->GetLongField(m_usage, committed_fid);
+    }
+    return usage;
 }
